Fixed Responsiveness and Access Control

+++ actions/comment_act.php
@@ -7,38 +7,43 @@
     //Initialize session
     session_start();
 
-    //Checks if the commentform is submitted
-    if(isset($_POST['commentForm'])) {
-        //Get data from url using GET method
-        $postId = $_GET['post_id'];
-        //Get user_id from session
-        $userId = $_SESSION['user_id'];
-        //Get data from form
-        $comment = $_POST['comment'];
-        //Call addComment function
-        createComment($conn, $userId, $postId, $comment);
-    }
+    if(isset($_SESSION['user_id'])) {
+        //Checks if the commentform is submitted
+        if(isset($_POST['commentForm'])) {
+            //Get data from url using GET method
+            $postId = $_GET['post_id'];
+            //Get user_id from session
+            $userId = $_SESSION['user_id'];
+            //Get data from form
+            $comment = $_POST['comment'];
+            //Call addComment function
+            createComment($conn, $userId, $postId, $comment);
+        }
 
-    //Checks if the replyform is submitted
-    if(isset($_POST['replyForm'])) {
-        //Get data from url using GET method
-        $commentId = $_GET['comment_id'];
-        $postId = $_GET['post_id'];
-        //Get user_id from session
-        $userId = $_SESSION['user_id'];
-        //Get data from form
-        $content = $_POST['content'];
-        //Call addReply function
-        createReply($conn, $userId, $postId, $commentId, $content);
-    }
+        //Checks if the replyform is submitted
+        if(isset($_POST['replyForm'])) {
+            //Get data from url using GET method
+            $commentId = $_GET['comment_id'];
+            $postId = $_GET['post_id'];
+            //Get user_id from session
+            $userId = $_SESSION['user_id'];
+            //Get data from form
+            $content = $_POST['content'];
+            //Call addReply function
+            createReply($conn, $userId, $postId, $commentId, $content);
+        }
 
-    //Check if the delete comment is submitted
-    if(isset($_POST['deleteComment'])) {
-        //Get data from url using GET method
-        $postId = $_GET['post_id'];
-        $commentId = $_GET['comment_id'];
+        //Check if the delete comment is submitted
+        if(isset($_POST['deleteComment'])) {
+            //Get data from url using GET method
+            $postId = $_GET['post_id'];
+            $commentId = $_GET['comment_id'];
 
-        //Call deleteComment function
-        deleteComment($conn, $commentId, $postId);
+            //Call deleteComment function
+            deleteComment($conn, $commentId, $postId);
+        }
+
+    } else {
+        header("Location: ../auth/login.php");
     }
 ?>
